# feat: Complete Phase 2 - FIX Protocol Integration, Smart Order Routing & Performance Benchmarking