--- a/docs/architecture/system_overview.md
+++ b/docs/architecture/system_overview.md
@@ -133,7 +133,7 @@ Threads communicate using lock-free queues to minimize contention.
 
 ## Current Implementation Status
 
-### Phase 1 (Completed) ✅
+### Phase 1 (Completed) 
 
 - Complete Order Book implementation with price level management
 - Basic Market Making strategy with dynamic spread adjustment
@@ -141,7 +141,7 @@ Threads communicate using lock-free queues to minimize contention.
 - Thread-safe data structures and utilities
 - Comprehensive unit tests
 
-### Phase 2 (Completed) ✅
+### Phase 2 (Completed) 
 
 - Lock-free data structures for all critical paths
 - Memory-mapped file system for data persistence