… CMakeLists.txt
… README.md
… core/routing/OrderRouter.cpp
… docs/FIX_INTEGRATION_SUMMARY.md
… docs/FIX_PROTOCOL_INTEGRATION.md
… docs/IB_TESTING_GUIDE.md
… docs/ORDER_ROUTING.md
… docs/PERFORMANCE_BENCHMARKS.md
… docs/PHASE_2_COMPLETION.md
… docs/ROADMAP.md
… docs/TESTING_GUIDE.md
… exchange/fix/FixConnector.cpp
… exchange/fix/InteractiveBrokersFixConnector.cpp
… exchange/fix/InteractiveBrokersFixConnector.h
… run-docker.sh
… run-native.sh
… tests/fix_basic_test.cpp
… tests/fix_message_test.cpp
… tests/performance/OrderRoutingBenchmark.cpp
… tests/routing_test.cpp